+++ rviz_mesh_map_plugin/include/rviz_mesh_map_plugin/MapDisplay.hpp
@@ -49,8 +49,8 @@
 #ifndef MAP_DISPLAY_HPP
 #define MAP_DISPLAY_HPP
 
-#include <Types.hpp>
-#include "RvizFileProperty.hpp"
+#include <rviz_mesh_map_plugin/Types.hpp>
+#include <rviz_mesh_map_plugin/RvizFileProperty.hpp>
 
 #include <vector>
 #include <memory>
@@ -69,8 +69,6 @@
 #include <QIcon>
 #include <QString>
 
-// #include <ros/ros.h>
-// #include <ros/console.h>
 #include <rviz_common/viewport_mouse_event.hpp>
 #include <rviz_common/visualization_manager.hpp>
 #include <rviz_common/visualization_frame.hpp>
@@ -98,20 +96,20 @@
 #ifndef Q_MOC_RUN
 #include <rviz_rendering/mesh_loader.hpp>
 
-#include <OGRE/OgreManualObject.h>
-#include <OGRE/OgreSceneNode.h>
-#include <OGRE/OgreSceneManager.h>
-#include <OGRE/OgreEntity.h>
-#include <OGRE/OgreStringConverter.h>
-#include <OGRE/OgreMaterialManager.h>
-#include <OGRE/OgreRay.h>
-#include <OGRE/OgreSceneQuery.h>
-#include <OGRE/OgreColourValue.h>
+#include <OgreManualObject.h>
+#include <OgreSceneNode.h>
+#include <OgreSceneManager.h>
+#include <OgreEntity.h>
+#include <OgreStringConverter.h>
+#include <OgreMaterialManager.h>
+#include <OgreRay.h>
+#include <OgreSceneQuery.h>
+#include <OgreColourValue.h>
 
 #endif
 
-#include <ClusterLabelDisplay.hpp>
-#include <MeshDisplay.hpp>
+#include <rviz_mesh_map_plugin/ClusterLabelDisplay.hpp>
+#include <rviz_mesh_map_plugin/MeshDisplay.hpp>
 
 namespace rviz
 {
@@ -125,7 +123,7 @@ class StringProperty;
 
 }  // End namespace rviz
 
-namespace rviz_map_plugin
+namespace rviz_mesh_map_plugin
 {
 using std::shared_ptr;
 using std::string;
@@ -222,9 +220,9 @@ private:
   std::string m_map_file_loaded;
 
   /// Subdisplay: ClusterLabel (for showing the clusters)
-  rviz_map_plugin::ClusterLabelDisplay* m_clusterLabelDisplay;
+  rviz_mesh_map_plugin::ClusterLabelDisplay* m_clusterLabelDisplay;
   /// Subdisplay: MeshDisplay (for showing the mesh)
-  rviz_map_plugin::MeshDisplay* m_meshDisplay;
+  rviz_mesh_map_plugin::MeshDisplay* m_meshDisplay;
 
   /**
    * @brief Create a RViz display from it's unique class_id
@@ -234,6 +232,6 @@ private:
   rviz_common::Display* createDisplay(const QString& class_id);
 };
 
-}  // end namespace rviz_map_plugin
+}  // end namespace rviz_mesh_map_plugin
 
 #endif